+++ src/gemseo_umdo/formulations/_functions/statistic_function_for_standard_sampling.py
@@ -21,15 +21,15 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from gemseo_umdo.formulations._functions.statistic_function_for_sampling import (
-    StatisticFunctionForSampling,
+from gemseo_umdo.formulations._functions.base_statistic_function_for_sampling import (
+    BaseStatisticFunctionForSampling,
 )
 
 if TYPE_CHECKING:
     from gemseo.typing import RealArray
 
 
-class StatisticFunctionForStandardSampling(StatisticFunctionForSampling):
+class StatisticFunctionForStandardSampling(BaseStatisticFunctionForSampling):
     """A function to compute a statistic from `Sampling`."""
 
     def _compute_statistic_estimation(