+++ src/gemseo_umdo/formulations/_statistics/sampling/factory.py
@@ -12,12 +12,19 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with this program; if not, write to the Free Software Foundation,
 # Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
-"""The base pilot class for the MLMC-MLCV algorithm."""
+"""A factory of statistic estimators for sampling-based U-MDO formulations."""
 
 from __future__ import annotations
 
-from gemseo_umdo.statistics.multilevel.pilot import Pilot
+from gemseo.core.base_factory import BaseFactory
+
+from gemseo_umdo.formulations._statistics.sampling.base_sampling_estimator import (
+    BaseSamplingEstimator,
+)
 
 
-class MLMCMLCVPilot(Pilot):
-    """The base pilot class for the MLMC-MLCV algorithm."""
+class SamplingEstimatorFactory(BaseFactory):
+    """The factory of sampling-based statistic estimators."""
+
+    _CLASS = BaseSamplingEstimator
+    _MODULE_NAMES = ("gemseo_umdo.formulations._statistics.sampling",)