+++ src/gemseo_umdo/scenarios/umdo_scenario.py
@@ -18,8 +18,8 @@ from __future__ import annotations
 
 from gemseo.scenarios.mdo_scenario import MDOScenario
 
-from gemseo_umdo.scenarios._scenario import _UScenario
+from gemseo_umdo.scenarios.base_u_scenario import BaseUScenario
 
 
-class UMDOScenario(_UScenario, MDOScenario):
+class UMDOScenario(BaseUScenario, MDOScenario):
     """An optimizer-based scenario for multidisciplinary design under uncertainty."""